feat(frontend): replace plugin task form with dialogs
This commit replaces the task forms on the PluginFile page with dialogs.

It also introduces the following improvements:
- When the chips overflow, it should push the content down and not overlap
- Entrypoint artifact param dialog wasn't resetting when closing and opening
- Make the dialogs persistent to avoid accidental closure by clicking outside of the dialog
- Renames the text "params" to "parameters"
